refactor: meetingHeroSection 서버 컴포넌트 전환 및 meetingHeroActions 분리
LCP 최적화를 위해 이미지 렌더링을 서버 컴포넌트로 이동.
클라이언트 인터랙션 로직(mutations, modals)은 meetingHeroActions로 분리.
